Move `StorageWithOverrides` to API server
