+++ core/lib/state/src/lib.rs
@@ -29,7 +29,6 @@ pub use self::{
     },
     shadow_storage::ShadowStorage,
     storage_factory::{BatchDiff, PgOrRocksdbStorage, ReadStorageFactory, RocksdbWithMemory},
-    storage_overrides::StorageWithOverrides,
     storage_view::{StorageView, StorageViewCache, StorageViewMetrics},
     witness::WitnessStorage,
 };
@@ -41,7 +40,6 @@ mod postgres;
 mod rocksdb;
 mod shadow_storage;
 mod storage_factory;
-mod storage_overrides;
 mod storage_view;
 #[cfg(test)]
 mod test_utils;
