+++ core/node/api_server/src/execution_sandbox/mod.rs
@@ -26,6 +26,7 @@ use super::tx_sender::MultiVMBaseSystemContracts;
 mod apply;
 mod error;
 mod execute;
+mod storage;
 pub mod testonly;
 #[cfg(test)]
 mod tests;
